+++ Expanded Cheat Menu/Changelog.md
@@ -14,6 +14,7 @@
 
 ### Fixed:
 - Newly spawned service buildings will have the settings from Service Building Stats properly applied (thanks hchsiao).
+- Upgraded buildings may reset some cheat settings (thanks hchsiao).
 
 ## v8.7 (04 Sep 2018)
 ### Changed: